+++ client/src/configuration.rs
@@ -107,7 +107,7 @@ build_config! {
         // Genesis section.
         (adaptive_weight_beta, (u64), ADAPTIVE_WEIGHT_DEFAULT_BETA)
         (anticone_penalty_ratio, (u64), ANTICONE_PENALTY_RATIO)
-        (chain_id, (Option<u64>), None)
+        (chain_id, (Option<u32>), None)
         // Snapshot Epoch Count is a consensus parameter. This flag overrides
         // the parameter, which only take effect in `dev` mode.
         (dev_snapshot_epoch_count, (u32), SNAPSHOT_EPOCHS_CAPACITY)
@@ -168,7 +168,7 @@ build_config! {
         (max_handshakes, (usize), 64)
         (max_incoming_peers, (usize), 64)
         (max_inflight_request_count, (u64), 64)
-        (max_outgoing_peers, (usize), 16)
+        (max_outgoing_peers, (usize), 8)
         (max_outgoing_peers_archive, (Option<usize>), None)
         (max_peers_tx_propagation, (usize), 128)
         (max_unprocessed_block_size_mb, (usize), (128))
@@ -196,11 +196,11 @@ build_config! {
         (node_table_timeout_s, (u64), 300)
         (node_table_promotion_timeout_s, (u64), 3 * 24 * 3600)
         (session_ip_limits, (String), "1,8,4,2".into())
-        (subnet_quota, (usize), 32)
+        (subnet_quota, (usize), 128)
 
         // Transaction cache/transaction pool section.
         (tx_cache_index_maintain_timeout_ms, (u64), 300_000)
-        (tx_pool_size, (usize), 500_000)
+        (tx_pool_size, (usize), 200_000)
         (tx_pool_min_tx_gas_price, (u64), 1)
         (tx_weight_scaling, (u64), 1)
         (tx_weight_exp, (u8), 1)
@@ -294,7 +294,7 @@ impl Configuration {
             Some(x) => x,
             // The default network id is 1 for historic reason. It doesn't
             // really matter.
-            None => self.raw_conf.chain_id.unwrap_or(1),
+            None => self.raw_conf.chain_id.unwrap_or(1) as u64,
         }
     }