+++ client/src/rpc/traits/mod.rs
@@ -2,12 +2,12 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-pub mod cfx;
-pub mod debug;
-pub mod pubsub;
-pub mod test;
-
 pub use cfx::Cfx;
 pub use debug::LocalRpc;
 pub use pubsub::PubSub;
 pub use test::TestRpc;
+
+pub mod cfx;
+pub mod debug;
+pub mod pubsub;
+pub mod test;