+++ client/src/rpc/types/block.rs
@@ -2,14 +2,21 @@
 // Conflux is free software and distributed under GNU General Public License.
 // See http://www.gnu.org/licenses/
 
-use crate::rpc::types::{Receipt, Transaction, H160, H256, U256};
-use cfx_types::U256 as CfxU256;
+use std::sync::Arc;
+
+use jsonrpc_core::Error as RpcError;
+use serde::{
+    de::{Deserialize, Deserializer, Error, Unexpected},
+    Serialize, Serializer,
+};
+use serde_json::Value;
+
+use cfx_types::{H160, H256, U256, U64};
 use cfxcore::{
     block_data_manager::{BlockDataManager, BlockExecutionResultWithEpoch},
     consensus::ConsensusGraphInner,
     SharedConsensusGraph,
 };
-use jsonrpc_core::Error as RpcError;
 use primitives::{
     receipt::{
         TRANSACTION_OUTCOME_EXCEPTION_WITHOUT_NONCE_BUMPING,
@@ -19,12 +26,8 @@ use primitives::{
     Block as PrimitiveBlock, BlockHeader as PrimitiveBlockHeader,
     BlockHeaderBuilder, TransactionIndex,
 };
-use serde::{
-    de::{Deserialize, Deserializer, Error, Unexpected},
-    Serialize, Serializer,
-};
-use serde_json::Value;
-use std::sync::Arc;
+
+use crate::rpc::types::{Receipt, Transaction};
 
 #[derive(PartialEq, Debug)]
 pub enum BlockTransactions {
@@ -101,7 +104,7 @@ pub struct Block {
     /// state_root/receipts_root/logs_bloom_hash/blame are not correct.
     /// It acts as a vote to help light client determining the
     /// state_root/receipts_root/logs_bloom_hash are correct or not.
-    pub blame: u32,
+    pub blame: U64,
     /// Transactions root hash
     pub transactions_root: H256,
     /// Epoch number
@@ -151,7 +154,7 @@ impl Block {
                         .map(|(idx, tx)| {
                             let receipt = execution_result.block_receipts.receipts.get(idx).unwrap();
                             let prior_gas_used = if idx == 0 {
-                                 CfxU256::zero()
+                                 U256::zero()
                             } else {
                                 execution_result.block_receipts.receipts.get(idx - 1).unwrap().accumulated_gas_used
                             };
@@ -215,7 +218,7 @@ impl Block {
             deferred_logs_bloom_hash: H256::from(
                 b.block_header.deferred_logs_bloom_hash().clone(),
             ),
-            blame: b.block_header.blame(),
+            blame: U64::from(b.block_header.blame()),
             transactions_root: H256::from(
                 b.block_header.transactions_root().clone(),
             ),
@@ -258,7 +261,7 @@ impl Block {
                     .with_deferred_logs_bloom_hash(
                         self.deferred_logs_bloom_hash.into(),
                     )
-                    .with_blame(self.blame)
+                    .with_blame(self.blame.as_u32())
                     .with_difficulty(self.difficulty.into())
                     .with_adaptive(self.adaptive)
                     .with_gas_limit(self.gas_limit.into())
@@ -369,7 +372,8 @@ impl Header {
 #[cfg(test)]
 mod tests {
     use super::{Block, BlockTransactions, Header};
-    use crate::rpc::types::{Transaction, H160, H256, U256};
+    use crate::rpc::types::Transaction;
+    use cfx_types::{H160, H256, U256};
     use keccak_hash::KECCAK_EMPTY_LIST_RLP;
     use serde_json;
 
@@ -411,7 +415,7 @@ mod tests {
             deferred_state_root: Default::default(),
             deferred_receipts_root: KECCAK_EMPTY_LIST_RLP.into(),
             deferred_logs_bloom_hash: cfx_types::KECCAK_EMPTY_BLOOM.into(),
-            blame: 0,
+            blame: 0.into(),
             transactions_root: KECCAK_EMPTY_LIST_RLP.into(),
             epoch_number: None,
             gas_limit: U256::default(),
@@ -426,12 +430,12 @@ mod tests {
         };
         let serialized_block = serde_json::to_string(&block).unwrap();
 
-        assert_eq!(serialized_block, r#"{"hash":"0x0000000000000000000000000000000000000000000000000000000000000000","parentHash":"0x0000000000000000000000000000000000000000000000000000000000000000","height":"0x0","miner":"0x0000000000000000000000000000000000000000","deferredStateRoot":"0x0000000000000000000000000000000000000000000000000000000000000000","deferredReceiptsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","deferredLogsBloomHash":"0xd397b3b043d87fcd6fad1291ff0bfd16401c274896d8c63a923727f077b8e0b5","blame":0,"transactionsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","epochNumber":null,"gasLimit":"0x0","timestamp":"0x0","difficulty":"0x0","powQuality":null,"refereeHashes":[],"adaptive":false,"nonce":"0x0","transactions":[],"size":"0x45"}"#);
+        assert_eq!(serialized_block, r#"{"hash":"0x0000000000000000000000000000000000000000000000000000000000000000","parentHash":"0x0000000000000000000000000000000000000000000000000000000000000000","height":"0x0","miner":"0x0000000000000000000000000000000000000000","deferredStateRoot":"0x0000000000000000000000000000000000000000000000000000000000000000","deferredReceiptsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","deferredLogsBloomHash":"0xd397b3b043d87fcd6fad1291ff0bfd16401c274896d8c63a923727f077b8e0b5","blame":"0x0","transactionsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","epochNumber":null,"gasLimit":"0x0","timestamp":"0x0","difficulty":"0x0","powQuality":null,"refereeHashes":[],"adaptive":false,"nonce":"0x0","transactions":[],"size":"0x45"}"#);
     }
 
     #[test]
     fn test_deserialize_block() {
-        let serialized = r#"{"hash":"0x0000000000000000000000000000000000000000000000000000000000000000","parentHash":"0x0000000000000000000000000000000000000000000000000000000000000000","height":"0x0","miner":"0x0000000000000000000000000000000000000000","deferredStateRoot":"0x0000000000000000000000000000000000000000000000000000000000000000","deferredReceiptsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","deferredLogsBloomHash":"0xd397b3b043d87fcd6fad1291ff0bfd16401c274896d8c63a923727f077b8e0b5","blame":0,"transactionsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","epochNumber":"0x0","gasLimit":"0x0","timestamp":"0x0","difficulty":"0x0","refereeHashes":[],"stable":null,"adaptive":false,"nonce":"0x0","transactions":[],"size":"0x45"}"#;
+        let serialized = r#"{"hash":"0x0000000000000000000000000000000000000000000000000000000000000000","parentHash":"0x0000000000000000000000000000000000000000000000000000000000000000","height":"0x0","miner":"0x0000000000000000000000000000000000000000","deferredStateRoot":"0x0000000000000000000000000000000000000000000000000000000000000000","deferredReceiptsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","deferredLogsBloomHash":"0xd397b3b043d87fcd6fad1291ff0bfd16401c274896d8c63a923727f077b8e0b5","blame":"0x0","transactionsRoot":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","epochNumber":"0x0","gasLimit":"0x0","timestamp":"0x0","difficulty":"0x0","refereeHashes":[],"stable":null,"adaptive":false,"nonce":"0x0","transactions":[],"size":"0x45"}"#;
         let result_block = Block {
             hash: H256::default(),
             parent_hash: H256::default(),
@@ -440,7 +444,7 @@ mod tests {
             deferred_state_root: Default::default(),
             deferred_receipts_root: KECCAK_EMPTY_LIST_RLP.into(),
             deferred_logs_bloom_hash: cfx_types::KECCAK_EMPTY_BLOOM.into(),
-            blame: 0,
+            blame: 0.into(),
             transactions_root: KECCAK_EMPTY_LIST_RLP.into(),
             epoch_number: Some(0.into()),
             gas_limit: U256::default(),