+++ client/src/rpc/types/call_request.rs
@@ -3,11 +3,13 @@
 // See http://www.gnu.org/licenses/
 
 use crate::rpc::types::Bytes;
-use cfx_types::{address_util::AddressUtil, Address, H160, U256};
+use cfx_types::{address_util::AddressUtil, Address, H160, U256, U64};
 use primitives::{
     transaction::Action, SignedTransaction, Transaction as PrimitiveTransaction,
 };
 use std::cmp::min;
+
+// use serde_json::de::ParserNumber::U64;
 
 #[derive(Debug, Default, PartialEq, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
@@ -27,7 +29,7 @@ pub struct CallRequest {
     /// Nonce
     pub nonce: Option<U256>,
     /// StorageLimit
-    pub storage_limit: Option<u64>,
+    pub storage_limit: Option<U64>,
 }
 
 #[derive(Debug, Default, PartialEq, Deserialize, Serialize)]
@@ -51,7 +53,7 @@ pub struct CheckBalanceAgainstTransactionResponse {
 }
 
 pub fn sign_call(
-    epoch_height: u64, chain_id: u64, request: CallRequest,
+    epoch_height: u64, chain_id: u32, request: CallRequest,
 ) -> SignedTransaction {
     let max_gas = U256::from(500_000_000);
     let gas = min(request.gas.unwrap_or(max_gas), max_gas);
@@ -67,7 +69,10 @@ pub fn sign_call(
         gas,
         gas_price: request.gas_price.unwrap_or_default(),
         value: request.value.unwrap_or_default(),
-        storage_limit: request.storage_limit.unwrap_or(std::u64::MAX),
+        storage_limit: request
+            .storage_limit
+            .map(|v| v.as_u64())
+            .unwrap_or(std::u64::MAX),
         epoch_height,
         chain_id,
         data: request.data.unwrap_or_default().into_vec(),
@@ -77,11 +82,14 @@ pub fn sign_call(
 
 #[cfg(test)]
 mod tests {
-    use super::CallRequest;
-    use cfx_types::{H160, U256};
+    use std::str::FromStr;
+
     use rustc_hex::FromHex;
     use serde_json;
-    use std::str::FromStr;
+
+    use cfx_types::{H160, U256, U64};
+
+    use super::CallRequest;
 
     #[test]
     fn call_request_deserialize() {
@@ -92,7 +100,7 @@ mod tests {
             "gas":"0x2",
             "value":"0x3",
             "data":"0x123456",
-            "storageLimit":123,
+            "storageLimit":"0x7b",
             "nonce":"0x4"
         }"#;
         let deserialized: CallRequest = serde_json::from_str(s).unwrap();
@@ -106,7 +114,7 @@ mod tests {
                 gas: Some(U256::from(2)),
                 value: Some(U256::from(3)),
                 data: Some(vec![0x12, 0x34, 0x56].into()),
-                storage_limit: Some(123),
+                storage_limit: Some(U64::from_str("7b").unwrap()),
                 nonce: Some(U256::from(4)),
             }
         );
@@ -120,7 +128,7 @@ mod tests {
             "gas": "0x76c0",
             "gasPrice": "0x9184e72a000",
             "value": "0x9184e72a",
-            "storageLimit":53758687,
+            "storageLimit":"0x3344adf",
             "data": "0xd46e8dd67c5d32be8d46e8dd67c5d32be8058bb8eb970870f072445675058bb8eb970870f072445675"
         }"#;
         let deserialized: CallRequest = serde_json::from_str(s).unwrap();
@@ -131,7 +139,7 @@ mod tests {
             gas_price: Some(U256::from_str("9184e72a000").unwrap()),
             gas: Some(U256::from_str("76c0").unwrap()),
             value: Some(U256::from_str("9184e72a").unwrap()),
-            storage_limit: Some(53758687),
+            storage_limit: Some(U64::from_str("3344adf").unwrap()),
             data: Some("d46e8dd67c5d32be8d46e8dd67c5d32be8058bb8eb970870f072445675058bb8eb970870f072445675".from_hex().unwrap().into()),
             nonce: None
         });