+++ core/src/evm/interpreter/mod.rs
@@ -51,7 +51,9 @@ use crate::{
     },
 };
 use bit_set::BitSet;
-use cfx_types::{Address, BigEndianHash, H256, U256, U512};
+use cfx_types::{
+    address_util::AddressUtil, Address, BigEndianHash, H256, U256, U512,
+};
 use std::{cmp, convert::TryFrom, marker::PhantomData, mem, sync::Arc};
 
 const GASOMETER_PROOF: &str = "If gasometer is None, Err is immediately returned in step; this function is only called by step; qed";
@@ -898,8 +900,37 @@ impl<Cost: CostType> Interpreter<Cost> {
                 // clear return data buffer before creating new call frame.
                 self.return_data = ReturnData::empty();
 
-                let can_call =
-                    has_balance && context.depth() < context.spec().max_depth;
+                let valid_code_address = code_address.is_valid_address();
+
+                let recipient_address = match instruction {
+                    instructions::CALL | instructions::STATICCALL => {
+                        &code_address
+                    }
+                    instructions::CALLCODE | instructions::DELEGATECALL => {
+                        &self.params.address
+                    }
+                    _ => unreachable!(),
+                };
+                let not_reentrancy_attack =
+                    // If this message call is not recursive call and reentering another contract,
+                    // we regard it as reentrancy.
+                    if context.is_reentrancy(&self.params.address, recipient_address) {
+                        // If this message call is invoked by `transfer()` or `send()`,
+                        // formally, the call data is empty and available gas is no more than 2300,
+                        // the reentrancy protection will not be triggered.
+                        if in_size.is_zero() {
+                            call_gas <= Cost::from(context.spec().call_stipend)
+                        } else {
+                            false
+                        }
+                    } else {
+                        true
+                    };
+
+                let can_call = has_balance
+                    && context.depth() < context.spec().max_depth
+                    && valid_code_address
+                    && not_reentrancy_attack;
                 if !can_call {
                     self.stack.push(U256::zero());
                     return Ok(InstructionResult::UnusedGas(call_gas));
@@ -984,7 +1015,8 @@ impl<Cost: CostType> Interpreter<Cost> {
             }
             instructions::SUICIDE => {
                 let address = self.stack.pop_back();
-                context.suicide(&u256_to_address(&address))?;
+                let refund_address = u256_to_address(&address);
+                context.suicide(&refund_address)?;
                 return Ok(InstructionResult::StopExecution);
             }
             instructions::LOG0
@@ -1629,7 +1661,7 @@ mod tests {
 
     #[test]
     fn should_not_fail_on_tracing_mem() {
-        let code = "7feeffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff006000527faaffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffaa6020526000620f120660406000601773945304eb96065b2a98b57a48a06ae28d285a71b56101f4f1600055".from_hex().unwrap();
+        let code = "7feeffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff006000527faaffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffaa6020526000620f120660406000601773145304eb96065b2a98b57a48a06ae28d285a71b56101f4f1600055".from_hex().unwrap();
 
         let mut params = ActionParams::default();
         params.address = Address::from_low_u64_be(5);