+++ tests/light/log_filtering_test.py
@@ -141,7 +141,7 @@ class LogFilteringTest(ConfluxTestFramework):
         tx = self.rpc[FULLNODE0].new_contract_tx(receiver="", data_hex=data_hex, sender=sender, priv_key=priv_key, storage_limit=1000)
         assert_equal(self.rpc[FULLNODE0].send_tx(tx, True), tx.hash_hex())
         receipt = self.rpc[FULLNODE0].get_transaction_receipt(tx.hash_hex())
-        assert_equal(receipt["outcomeStatus"], 0)
+        assert_equal(receipt["outcomeStatus"], "0x0")
         address = receipt["contractCreated"]
         assert_is_hex_string(address)
         return receipt, address
@@ -150,7 +150,7 @@ class LogFilteringTest(ConfluxTestFramework):
         tx = self.rpc[FULLNODE0].new_contract_tx(receiver=contract, data_hex=data_hex, sender=sender, priv_key=priv_key, storage_limit=1000)
         assert_equal(self.rpc[FULLNODE0].send_tx(tx, True), tx.hash_hex())
         receipt = self.rpc[FULLNODE0].get_transaction_receipt(tx.hash_hex())
-        assert_equal(receipt["outcomeStatus"], 0)
+        assert_equal(receipt["outcomeStatus"], "0x0")
         return receipt
 
 if __name__ == "__main__":